Alterações devido a erros durante o "loading" do banco de dados.

+++ src/main/java/br/com/grupomult/repositories/TipoCarroRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.grupomult.entities.TipoCarro;
 
-public interface TipoCarroRepository extends CrudRepository<TipoCarro, Integer>{
+public interface TipoCarroRepository extends CrudRepository<TipoCarro, Long>{
 	
 	public TipoCarro findByDescricao(String descricao);
